requirements.txt
pyqt6
pyqt6-sip
pyserial